+++ packages/werkbank/lib/src/addons/src/knobs/src/knobs/list_knob.dart
@@ -26,7 +26,7 @@ extension ListKnobExtension on KnobsComposer {
   }
 }
 
-extension NullableListKnobExtension on NullableKnobs {
+extension NullableListKnobExtension on NullableKnobsComposer {
   WritableKnob<T?> list<T extends Object>(
     String label, {
     required T initialValue,